requirements.txt
allure-pytest==2.13.3
allure-python-commons==2.13.3
assertpy==1.1
atomicwrites==1.4.1
attrs==23.1.0
bcrypt==4.0.1
build==1.1.1
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
curlify==2.2.1
distlib==0.3.7
exceptiongroup==1.2.0
execnet==2.0.2
Faker==15.1.5
filelock==3.13.1
h11==0.14.0
identify==2.5.32
idna==3.6
imap-tools==1.5.0
iniconfig==2.0.0
nodeenv==1.8.0
outcome==1.3.0.post0
packaging==23.2
paramiko==3.4.0
pip-tools==7.4.0
pip-upgrade==0.0.6
platformdirs==4.0.0
pluggy==1.3.0
postgres==4.0
pre-commit==3.5.0
psycopg2-binary==2.9.6
psycopg2-pool==1.1
py==1.11.0
pycparser==2.21
PyHamcrest==2.0.4
PyJWT==2.8.0
PyNaCl==1.5.0
pyproject_hooks==1.0.0
PySocks==1.7.1
pytest==7.1.2
pytest-rerunfailures==13.0
pytest-xdist==3.5.0
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
sshtunnel==0.4.0
tomli==2.0.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.1.0
virtualenv==20.24.7
wsproto==1.2.0
